* commit '56f39a315c213f51b86e89f4c6e2c3f206fb209f':
  [FIX]fix a merge
  [ADD]add a new sql file
  [UPDATE]update funcionality and add automatic propertie

+++ dataset-service/src/main/java/org/eea/dataset/service/impl/DatasetSnapshotServiceImpl.java
@@ -12,6 +12,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import org.bson.types.ObjectId;
+import org.eea.dataset.controller.DataSetSnapshotControllerImpl;
 import org.eea.dataset.mapper.ReleaseMapper;
 import org.eea.dataset.mapper.SnapshotMapper;
 import org.eea.dataset.mapper.SnapshotSchemaMapper;
@@ -22,12 +23,14 @@ import org.eea.dataset.persistence.metabase.domain.DataSetMetabase;
 import org.eea.dataset.persistence.metabase.domain.DesignDataset;
 import org.eea.dataset.persistence.metabase.domain.EUDataset;
 import org.eea.dataset.persistence.metabase.domain.PartitionDataSetMetabase;
+import org.eea.dataset.persistence.metabase.domain.ReportingDataset;
 import org.eea.dataset.persistence.metabase.domain.Snapshot;
 import org.eea.dataset.persistence.metabase.domain.SnapshotSchema;
 import org.eea.dataset.persistence.metabase.repository.DataCollectionRepository;
 import org.eea.dataset.persistence.metabase.repository.DataSetMetabaseRepository;
 import org.eea.dataset.persistence.metabase.repository.EUDatasetRepository;
 import org.eea.dataset.persistence.metabase.repository.PartitionDataSetMetabaseRepository;
+import org.eea.dataset.persistence.metabase.repository.ReportingDatasetRepository;
 import org.eea.dataset.persistence.metabase.repository.SnapshotRepository;
 import org.eea.dataset.persistence.metabase.repository.SnapshotSchemaRepository;
 import org.eea.dataset.persistence.schemas.domain.DataSetSchema;
@@ -49,9 +52,11 @@ import org.eea.interfaces.controller.document.DocumentController.DocumentControl
 import org.eea.interfaces.controller.recordstore.RecordStoreController.RecordStoreControllerZuul;
 import org.eea.interfaces.controller.ums.UserManagementController.UserManagementControllerZull;
 import org.eea.interfaces.controller.validation.RulesController.RulesControllerZuul;
+import org.eea.interfaces.controller.validation.ValidationController.ValidationControllerZuul;
 import org.eea.interfaces.vo.dataflow.DataFlowVO;
 import org.eea.interfaces.vo.dataflow.DataProviderVO;
 import org.eea.interfaces.vo.dataflow.RepresentativeVO;
+import org.eea.interfaces.vo.dataset.CreateSnapshotVO;
 import org.eea.interfaces.vo.dataset.DataSetMetabaseVO;
 import org.eea.interfaces.vo.dataset.enums.DatasetTypeEnum;
 import org.eea.interfaces.vo.dataset.enums.ErrorTypeEnum;
@@ -131,6 +136,9 @@ public class DatasetSnapshotServiceImpl implements DatasetSnapshotService {
   @Autowired
   private ValidationRepository validationRepository;
 
+  @Autowired
+  private ReportingDatasetRepository reportingDatasetRepository;
+
   /** The dataset service. */
   @Autowired
   @Qualifier("proxyDatasetService")
@@ -187,6 +195,12 @@ public class DatasetSnapshotServiceImpl implements DatasetSnapshotService {
   /** The user management controller zull. */
   @Autowired
   private UserManagementControllerZull userManagementControllerZull;
+
+  @Autowired
+  private ValidationControllerZuul validationControllerZuul;
+
+  @Autowired
+  private DataSetSnapshotControllerImpl dataSetSnapshotControllerImpl;
 
   /** The Constant FILE_PATTERN_NAME. */
   private static final String FILE_PATTERN_NAME = "schemaSnapshot_%s-DesignDataset_%s";
@@ -259,7 +273,7 @@ public class DatasetSnapshotServiceImpl implements DatasetSnapshotService {
    */
   @Override
   @Async
-  public void addSnapshot(Long idDataset, String description, Boolean released,
+  public void addSnapshot(Long idDataset, CreateSnapshotVO createSnapshotVO,
       Long partitionIdDestination) {
 
 
@@ -271,7 +285,7 @@ public class DatasetSnapshotServiceImpl implements DatasetSnapshotService {
       // 1. Create the snapshot in the metabase
       Snapshot snap = new Snapshot();
       snap.setCreationDate(java.sql.Timestamp.valueOf(LocalDateTime.now()));
-      snap.setDescription(description);
+      snap.setDescription(createSnapshotVO.getDescription());
       DataSetMetabase dataset = new DataSetMetabase();
       dataset.setId(idDataset);
       if (DatasetTypeEnum.REPORTING.equals(datasetService.getDatasetType(idDataset))) {
@@ -284,9 +298,13 @@ public class DatasetSnapshotServiceImpl implements DatasetSnapshotService {
       }
       snap.setReportingDataset(dataset);
       snap.setDataSetName("snapshot from dataset_" + idDataset);
-      snap.setDcReleased(released);
+      snap.setDcReleased(createSnapshotVO.getReleased());
       snap.setEuReleased(false);
       snap.setBlocked(isBlocked != null && !isBlocked.isEmpty());
+
+      snap.setAutomatic(
+          Boolean.TRUE.equals(createSnapshotVO.getAutomatic()) ? Boolean.TRUE : Boolean.FALSE);
+
       snapshotRepository.save(snap);
       LOG.info("Snapshot {} created into the metabase", snap.getId());
       snap.getId();
@@ -312,7 +330,7 @@ public class DatasetSnapshotServiceImpl implements DatasetSnapshotService {
       List<Object> criteria = new ArrayList<>();
       criteria.add(LockSignature.CREATE_SNAPSHOT.getValue());
       criteria.add(idDataset);
-      criteria.add(released);
+      criteria.add(createSnapshotVO.getReleased());
       lockService.removeLockByCriteria(criteria);
     }
     // release snapshot when the user press create+release
@@ -351,6 +369,11 @@ public class DatasetSnapshotServiceImpl implements DatasetSnapshotService {
    */
   @Override
   public void removeSnapshot(Long idDataset, Long idSnapshot) throws EEAException {
+
+    if (Boolean.TRUE.equals(snapshotRepository.findById(idSnapshot).get().getAutomatic())) {
+      LOG_ERROR.error("Error deleting snapshot, the snapshot is automatic");
+      throw new EEAException(EEAErrorMessage.ERROR_DELETING_SNAPSHOT);
+    }
     // Remove from the metabase
     snapshotRepository.deleteById(idSnapshot);
     // Delete the file
@@ -752,29 +775,6 @@ public class DatasetSnapshotServiceImpl implements DatasetSnapshotService {
     });
   }
 
-  /**
-   * Delete all snapshots.
-   *
-   * @param idDataset the id dataset
-   *
-   * @throws EEAException the EEA exception
-   */
-  @Override
-  @Async
-  public void deleteAllSnapshots(Long idDataset) throws EEAException {
-
-    LOG.info("Deleting all snapshots when the dataset it's going to be deleted");
-    List<SnapshotVO> snapshots = getSnapshotsByIdDataset(idDataset);
-    snapshots.stream().forEach(s -> {
-      try {
-        removeSnapshot(idDataset, s.getId());
-      } catch (EEAException e) {
-        LOG_ERROR.error("Error deleting the snapshot with id {} due to reason {} ", s.getId(),
-            e.getMessage(), e);
-      }
-    });
-
-  }
 
   /**
    * Obtain partition.
@@ -985,6 +985,47 @@ public class DatasetSnapshotServiceImpl implements DatasetSnapshotService {
       }
     }
     return releases;
+  }
+
+
+  /**
+   * Creates the release snapshots.
+   *
+   * @param dataflowId the dataflow id
+   * @param dataProviderId the data provider id
+   */
+  @Override
+  @Async
+  public void createReleaseSnapshots(Long dataflowId, Long dataProviderId) {
+    List<Long> datasetsFilters = reportingDatasetRepository.findByDataflowId(dataflowId).stream()
+        .filter(rd -> rd.getDataProviderId().equals(dataProviderId)).map(ReportingDataset::getId)
+        .collect(Collectors.toList());
+
+    List<Boolean> oneHaveBlockers = new ArrayList();
+    // validationControllerZuul.validateDataSetDataFuture(datasetsFilters);
+    datasetsFilters.forEach(id -> {
+      TenantResolver.setTenantName(String.format(LiteralConstants.DATASET_FORMAT_NAME, id));
+      oneHaveBlockers.add(validationRepository.existsByLevelError(ErrorTypeEnum.BLOCKER));
+    });
+
+    if (oneHaveBlockers.contains(Boolean.TRUE)) {
+      System.err.println("ERROR");
+    } else {
+      // datasetsFilters.forEach(id -> {
+      // CreateSnapshotVO createSnapshotVO = new CreateSnapshotVO();
+      // createSnapshotVO.setReleased(true);
+      // createSnapshotVO.setAutomatic(Boolean.TRUE);
+      // Date ahora = new Date();
+      // SimpleDateFormat formateador = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+      // createSnapshotVO.setDescription("Release " + formateador.format(ahora).toString());
+      // dataSetSnapshotControllerImpl.createSnapshot(id, createSnapshotVO);
+      // });
+    }
+    List<Object> criteria = new ArrayList<>();
+    criteria.add(LockSignature.RELEASE_SNAPSHOTS.getValue());
+    criteria.add(dataflowId);
+    criteria.add(dataProviderId);
+    lockService.removeLockByCriteria(criteria);
   }
 
 }
